+++ changelog.md
@@ -4,6 +4,8 @@
 
 - Disabled the enhanced editor toggle button by default.
 - Disabled the line break indicator by default.
+- Added an option to disable saving when exiting using the 'Escape' key.
+- Added a 'Revert' GUI button.
 
 ## 1.2.0